Fully implemented suggestion logging and approval
When a suggestion is made it'll be added to a .json store acting as an array database. The ID that is stored is just the Discord message ID, so when the suggestion is approved it'll be removed from the datastore.

This allows for the bot to restart without forgetting previously generated suggestions.